Initial commit of half written project. This project will find memory leaks in application.
Wrappers over memalloc, free and others are already working. Data are stored
in to logging file each time malloc and free are called. This behavior will be
optimised in future.

Test application is also half finished but same memory leaks need to be added.

Data post processing is not finished at all, but it will be soon.