Add support for all chunk types in to post process application.
All chunk types are basically more complicated version of malloc and free chunks. For example realloc is not nothing else than free and malloc combined.
Memalign could be changed into malloc chunk type in future. Current form of memalign implies from not understanding whole problematic in time of design.

The calloc and realloc are sended in this form to reduce load on analysed CPU.